Upload works
Uploading works!!!!!

I just need some help changing the database values. you would need to
store the description, update the value hasUploaded, and change the date
so it is when it was uploaded. But it does store it in a folder called
resume. You might have to make it, not sure if it will make it.
Once that is done we can make the resumeeditor page. Send it the
username of the person loggedin with ?pdf=uwinid

Shouldnt be to bad.